requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
anthropic>=0.25.0
openai-whisper>=20231117
httpx>=0.27.0
python-multipart>=0.0.9
aiofiles>=23.2.1
aiosmtplib>=3.0.0
email-validator>=2.1.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
apscheduler>=3.10.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
